Feat(Stats): Fixed enemy and tower stats/display

+++ src/entities/enemies/Enemy.h
@@ -34,8 +34,8 @@ protected:
    
     // Otros atributos
     std::string enemyType;
-    int health;
-    int speed;
+    double health;
+    double speed;
     int arrowResistance;
     int magicResistance;
     int artilleryResistance;
@@ -45,7 +45,7 @@ protected:
 
 public:
     Enemy(bool alive, Vector2 position, int frameSpeed, const std::string& texturePath, int frameCount,
-          const std::string& enemyType, int health, int speed, int arrowResistance, int magicResistance,
+          const std::string& enemyType, double health, double speed, int arrowResistance, int magicResistance,
           int artilleryResistance, bool mutated, int generation, int mutationChance);
     virtual ~Enemy();
 
@@ -54,10 +54,10 @@ public:
     void SetEnemyType(const std::string& type);
 
     int GetHealth() const;
-    void SetHealth(int health);
+    void SetHealth(double health);
 
     int GetSpeed() const;
-    void SetSpeed(int speed);
+    void SetSpeed(double speed);
 
     int GetArrowResistance() const;
     void SetArrowResistance(int resistance);
